# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASHPB PV ESS Visualization API",
    "description": "API to serve downsampled simulation data for visualization.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/simulation/variables": {
      "get": {
        "tags": [
          "Simulation"
        ],
        "summary": "Get Variables",
        "description": "Get all available variables and their units for plotting.",
        "operationId": "simulation_get_variables",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Simulation Get Variables"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulation/results": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Get Results",
        "description": "Fetch downsampled results using LTTB algorithm.",
        "operationId": "simulation_get_results",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownsampledResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Read Root",
        "operationId": "read_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DownsampledResultRequest": {
        "properties": {
          "variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variables"
          },
          "threshold": {
            "type": "integer",
            "title": "Threshold",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "variables"
        ],
        "title": "DownsampledResultRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}